requirements.txt
alembic==1.10.2
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
SQLAlchemy==2.0.7
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
gunicorn==20.1.0
redis==4.5.5
alembic==1.11.1
python-dotenv==1.0.0


